agent/pyproject.toml
[tool.poetry]
name = "rmirror-agent"
version = "0.1.0"
description = "Mac background service for syncing reMarkable tablets to rMirror Cloud"
authors = ["rMirror Cloud <support@rmirror.cloud>"]
license = "AGPL-3.0"
readme = "README.md"
packages = [{include = "app"}]

[tool.poetry.dependencies]
python = "^3.11,<3.14"
watchdog = "^4.0.0"  # File system monitoring
httpx = "^0.27.0"  # Async HTTP client
flask = "^3.0.0"  # Web UI server
pyyaml = "^6.0.1"  # Configuration parsing
pydantic = "^2.5.0"  # Configuration validation
pydantic-settings = "^2.1.0"  # Settings management
rumps = "^0.4.0"  # macOS menu bar app (Mac only)
click = "^8.1.7"  # CLI interface

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
pytest-asyncio = "^0.23.2"
mypy = "^1.7.1"
ruff = "^0.1.7"
types-pyyaml = "^6.0.12"

[tool.poetry.scripts]
rmirror-agent = "app.main:cli"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]  # Line too long (handled by formatter)

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
plugins = ["pydantic.mypy"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
asyncio_mode = "auto"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
